requirements.txt
flask
psycopg2-binary
python-dotenv
flask-mail
gunicorn
